add fix

+++ lib/data/provider/collections_provider.dart
@@ -78,7 +78,7 @@ class CollectionsProvider extends ChangeNotifier {
     for (final entry in _collections.entries) {
       entry.value.data.removeWhere((a) => a.uniqueId == uniqueId);
     }
-    if (kodikResult != null) {
+    if (kodikResult != null && anime.release.shikimoriId != null) {
       await _repository.addToCollection(
         type,
         int.parse('$kodikIdPattern${anime.release.shikimoriId}'),
@@ -125,7 +125,7 @@ class CollectionsProvider extends ChangeNotifier {
       entry.value.data.removeWhere((a) => a.uniqueId == uniqueId);
     }
 
-    if (kodikResult != null) {
+    if (kodikResult != null && anime.release.shikimoriId != null) {
       await _repository.removeFromCollection(
         type,
         int.parse('$kodikIdPattern${anime.release.shikimoriId}'),
